pyproject.toml
[project]
name = "events-agent"
version = "0.1.0"
description = "An events agent using OpenAI API"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.1",
    "openai-agents>=0.0.9",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 
